04_Curso/Informática_Gráfica_Visualización/lab-1/requirements.txt
glfw==2.7.0
numpy==2.1.2
PyOpenGL==3.1.7
tk==0.1.0
